% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cd_related_codes.R
\name{calc_A_est_by_lam_cd}
\alias{calc_A_est_by_lam_cd}
\title{Calculate an Adjacency Matrix using sparsebn object}
\usage{
calc_A_est_by_lam_cd(result_cd)
}
\arguments{
\item{result_cd}{blah}
}
\value{
blah
}
\description{
Calculate an Adjacency Matrix using sparsebn object
}
