% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_eval.R
\name{mod_sel}
\alias{mod_sel}
\title{Select Model using the result from eval_by_lam}
\usage{
mod_sel(one_simu_result, n_edges, sel_crit = "JI")
}
\arguments{
\item{one_simu_result}{blah}

\item{n_edges}{blah}

\item{sel_crit}{blah}
}
\value{
blah blah blah
}
\description{
Select Model using the result from eval_by_lam
}
